+++ cfgov/unprocessed/apps/owning-a-home/css/main.less
@@ -17,3 +17,4 @@
 @import (less) "./form-resources.less";
 @import (less) "./form-field-with-button.less";
 @import (less) "./no-js.less";
+@import (less) "./tab.less";
