+++ cfgov/unprocessed/js/routes/on-demand/form-explainer.js
@@ -2,6 +2,6 @@
    Scripts for Form Explainer
    ========================================================================== */
 
-import formExplainer from '../../apps/owning-a-home/form-explainer';
+import formExplainer from '../../../apps/owning-a-home/js/form-explainer';
 
 formExplainer.init();
